RootScout/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx>=0.28.1,<1.0.0
python-dotenv==1.0.1
opentelemetry-proto
